See previous commit message
I again had an issue with git becoming screwy.  It has something
to do with admin privileges in my C: directory, but I'm not exactly
sure what since it seems to work sometimes and not other times.  I
ended up having to run bash as an admin and then re-add and commit.